% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allda.R
\name{allda}
\alias{allda}
\title{Adaptive Local Linear Discriminant Analysis (ALLDA)}
\usage{
allda(
  X,
  y,
  ncomp,
  k,
  r = 2,
  preproc = multivarious::center(),
  max_iter = 30,
  tol = 1e-04,
  var_retained = 0.95,
  reg = 1e-05
)
}
\arguments{
\item{X}{Data matrix of size n x d (n samples, d features).}

\item{y}{Label vector of length n.}

\item{ncomp}{Reduced dimension (must be less than the number of features retained by PCA).}

\item{k}{Number of neighbors.}

\item{r}{Parameter r (default is 2, must be > 1).}

\item{preproc}{A preprocessing step from \code{multivarious}. Defaults to centering.}

\item{max_iter}{Maximum number of iterations (default is 30).}

\item{tol}{Convergence tolerance (default is 1e-4).}

\item{var_retained}{Proportion of variance to retain during PCA (default is 0.95).}

\item{reg}{Regularization term to ensure invertibility of St (default is 1e-5).}
}
\value{
An S3 object of class "discriminant_projector" containing the transformation matrix W,
the transformed scores, and related metadata.
}
\description{
Perform dimensionality reduction using the ALLDA algorithm.
}
\examples{
# result <- allda(X, y, ncomp = 2, k = 5)
# W <- result$rotation
# S <- result$S
}
\references{
Nie, F., Wang, Z., Wang, R., Wang, Z., & Li, X. (2020). Adaptive local linear discriminant analysis.
ACM Transactions on Knowledge Discovery from Data (TKDD), 14(1), 1-19.
}
