% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allda.R
\name{apply_pca}
\alias{apply_pca}
\title{Apply PCA for Null Space Removal}
\usage{
apply_pca(X, var_retained = 0.95)
}
\arguments{
\item{X}{Data matrix of size n x d (n samples, d features).}

\item{var_retained}{Proportion of variance to retain (default is 0.95).}
}
\value{
A list containing:
\item{X_pca}{The transformed data matrix after PCA (n x d_pca).}
\item{rot}{The PCA rotation/loadings matrix (d x d_pca).}
\item{fit}{The PCA fit object from \code{multivarious::pca}.}
}
\description{
Preprocess the data matrix to remove the null space using PCA
and retain a number of principal components that explain at least
\code{var_retained} proportion of variance.
}
