% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter.R
\name{between_class_scatter}
\alias{between_class_scatter}
\title{Calculate Between-Class Scatter Matrix (via C++)}
\usage{
between_class_scatter(X, Y)
}
\arguments{
\item{X}{A numeric matrix (n x d).}

\item{Y}{A factor or numeric vector of length n.}
}
\value{
A numeric (d x d) between-class scatter matrix.
}
\description{
Computes the between-class scatter matrix using a C++ function \code{betweenSScpp}.
}
\keyword{internal}
