% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cces}
\alias{cces}
\title{Cooperative Congressional Election Study 2018}
\format{
\subsection{\code{cces}}{

A data frame with 1,000 rows and 15 columns:
\describe{
\item{age}{Age (in years)}
\item{female}{Gender (1 = female)}
\item{educ_cont}{Education level (1-6)}
\item{pid_cont}{Party identification (1-7)}
\item{educ_pid}{educ_cont * pid_cont}
\item{oe01-oe10}{Open-ended responses}
}
}
}
\source{
\url{https://cces.gov.harvard.edu/}
}
\usage{
cces
}
\description{
A subset of data from the UWM Team Content of the 2018 CCES wave. See Kraft (2023) for details.
}
\keyword{datasets}
