% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpage.R
\name{compute_diversity_graph}
\alias{compute_diversity_graph}
\title{Compute the manual diversity graph D (Eq. (7) in the paper)}
\usage{
compute_diversity_graph(X, y)
}
\arguments{
\item{X}{A numeric matrix of size \eqn{n \times d}.}

\item{y}{A length-\eqn{n} vector/factor of class labels.}
}
\value{
An \eqn{n \times n} matrix \eqn{D}.
}
\description{
For each pair of samples \eqn{(x_i, x_j)}, the entry \eqn{d_{ij}} in the diversity
matrix \eqn{D} is:
\deqn{
  d_{ij} =
  \begin{cases}
    \frac{1}{2}\left( e^{\|x_i - x_j\|^2 / \rho_i^+} + e^{\|x_j - x_i\|^2 / \rho_j^+} \right), & \text{if } c_i = c_j, \\
    \frac{1}{2}\left( e^{\|x_i - x_j\|^2 / \rho_i^-} + e^{\|x_j - x_i\|^2 / \rho_j^-} \right), & \text{if } c_i \neq c_j,
  \end{cases}
}
where \eqn{\rho_i^+} and \eqn{\rho_i^-} follow Eqs. (5) and (6) \strong{but} with \eqn{q=0}.
}
