% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dict_sample}
\alias{dict_sample}
\title{Constraint Dictionary}
\format{
\subsection{\code{cces}}{

A data character vector with 4 elements:
\describe{
\item{conjunctions}{also, and}
\item{exclusive}{but, without}
}
}
}
\usage{
dict_sample
}
\description{
A sample of terms that signal a higher level of constraint between different considerations (combining conjunctions and exclusive words). See Kraft (2023) for details.
}
\keyword{datasets}
