% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discursive.R
\name{discursive}
\alias{discursive}
\title{Title}
\usage{
discursive(
  data,
  openends,
  dictionary,
  meta,
  customstopwords = NULL,
  lower.tresh = 10,
  K = 0,
  seed = NULL
)
}
\arguments{
\item{seed}{}
}
\description{
Title
}
