% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discursive_size.R
\name{discursive_size}
\alias{discursive_size}
\title{Title}
\usage{
discursive_size(
  data,
  openends,
  meta,
  customstopwords = NULL,
  lower.tresh = 10,
  K = 25,
  seed = 12345
)
}
\arguments{
\item{seed}{}
}
\description{
Title
}
