% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moda.R
\name{.approximate_covariance}
\alias{.approximate_covariance}
\title{Approximate Covariance as U_i Lambda_i U_i^T + sigma^2 I  (Section 6)}
\usage{
.approximate_covariance(Xc, energyThreshold = 0.9)
}
\description{
Approximate Covariance as U_i Lambda_i U_i^T + sigma^2 I  (Section 6)
}
\keyword{internal}
