% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moda.R
\name{.cluster_each_class}
\alias{.cluster_each_class}
\title{Cluster each class into multiple clusters}
\usage{
.cluster_each_class(X_work, y, numClusters, method)
}
\description{
Cluster each class into multiple clusters
}
\keyword{internal}
