% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moda.R
\name{.multiply_sigma_B}
\alias{.multiply_sigma_B}
\title{Multiply approximate covariance by B
Sigma_i = U_i Lambda_i U_i^T + sigma^2 I
=> Sigma_i B = U_i Lambda_i (U_i^T B) + sigma^2 B}
\usage{
.multiply_sigma_B(SigmaApprox, B)
}
\description{
Multiply approximate covariance by B
Sigma_i = U_i Lambda_i U_i^T + sigma^2 I
=> Sigma_i B = U_i Lambda_i (U_i^T B) + sigma^2 B
}
\keyword{internal}
