% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moda.R
\name{.safe_invert}
\alias{.safe_invert}
\title{Safely invert a matrix, fallback to pseudo-inverse if needed}
\usage{
.safe_invert(M)
}
\description{
Safely invert a matrix, fallback to pseudo-inverse if needed
}
\keyword{internal}
