% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discursive_size.R
\name{ntopics}
\alias{ntopics}
\title{Compute number of topics based on stm results}
\usage{
ntopics(x, docs)
}
\arguments{
\item{x:}{stm model result}

\item{docs:}{}
}
\value{
data.frame: estimated number of topics and distinctiveness of word choice
}
\description{
Compute number of topics based on stm results
}
