% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discursive_range.R
\name{oe_shannon}
\alias{oe_shannon}
\title{Compute Shannon entropy}
\usage{
oe_shannon(x)
}
\arguments{
\item{x}{Character vector containing open-ended responses}
}
\value{
Numeric vector with the same length as x. Contains Shannon entropy in individual response lengths re-scaled to range from 0 to 1.
}
\description{
Compute Shannon entropy
}
\examples{
\dontrun{
oe_shannon(c(paste(rep("Word", 10), collapse = " "), "Word"))
}
}
