% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/olda.R
\name{olda}
\alias{olda}
\title{Orthogonal Linear Discriminant Analysis}
\usage{
olda(X, Y, preproc = pass())
}
\arguments{
\item{X}{A numeric matrix where columns represent features and rows represent samples.}

\item{Y}{A factor or numeric vector representing class labels for each sample.}

\item{preproc}{A preprocessing function to apply to the data. Default is no preprocessing.}
}
\value{
An object of class \code{projector} containing the orthogonal projection matrix and additional information about the projection.
}
\description{
This function performs Orthogonal Linear Discriminant Analysis (OLDA) on a given dataset.
OLDA finds orthogonal linear discriminants that maximize the separation between classes.
It is useful for dimensionality reduction and classification tasks.
}
\details{
The function proceeds through the following steps:
\enumerate{
\item \strong{ULDA Projection}: The data is first projected using Uncorrelated Linear Discriminant Analysis (ULDA).
\item \strong{QR Decomposition}: The projection matrix from ULDA is then orthogonalized using QR decomposition.
\item \strong{Final Projection}: The orthogonalized projection matrix is returned.
}
}
\examples{
data(iris)
X <- as.matrix(iris[, 1:4])
Y <- iris[, 5]
res <- olda(X, Y)

}
\seealso{
\code{\link{ulda}}, \code{\link{qr}}, \code{\link{qr.Q}}
}
