% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpage.R
\name{predict.dgpage_projector}
\alias{predict.dgpage_projector}
\title{Predict method for a \code{dgpage_projector} object}
\usage{
\method{predict}{dgpage_projector}(
  object,
  new_data,
  method = c("1nn", "lda", "euclid"),
  type = c("class", "prob"),
  k = 1,
  ...
)
}
\arguments{
\item{object}{A \code{dgpage_projector} object.}

\item{new_data}{A numeric matrix (m x d).}

\item{method}{One of \code{c("1nn","lda","euclid")}. Default \code{"1nn"}.}

\item{type}{If \code{method="1nn"}, we only return \code{"class"}.
If \code{method \%in\% c("lda","euclid")}, we can do \code{"class"} or \code{"prob"}.}

\item{k}{Number of neighbors if using \code{method="1nn"}. Default 1 (which is actually 1NN).}

\item{...}{Not used (or pass \code{method="lda"} / \code{type="prob"}).}
}
\value{
If \code{method="1nn"}, a factor vector of length \code{m}.
If \code{method} is \code{"lda"} or \code{"euclid"} and \code{type="class"},
a factor of length \code{m}.
If \code{method="lda"} or \code{"euclid"} and \code{type="prob"}, an
\code{m x nclass} matrix of posterior-like values.
}
\description{
Extends the discriminant methods with a \code{method="1nn"} option that
uses \pkg{Rnanoflann} for 1-nearest-neighbor classification in the projected space.
}
\details{
Otherwise, if \code{method} is \code{"lda"} or \code{"euclid"}, we fall back to the
standard \code{\link{predict.discriminant_projector}} method for linear discriminant
or Euclidean nearest-mean classification in subspace.
}
\seealso{
\code{\link{predict.discriminant_projector}}
}
