% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shrinkda.R
\name{shrinkage_da}
\alias{shrinkage_da}
\title{data(iris)
X <- as.matrix(iris\link{,1:4})
Y <- iris\link{,5}
res <- shrinkda(X,Y)}
\usage{
shrinkage_da(
  X,
  Y,
  preproc = center(),
  lambda,
  lambda.var,
  lambda.freqs,
  diagonal = FALSE,
  verbose = FALSE
)
}
\description{
data(iris)
X <- as.matrix(iris\link{,1:4})
Y <- iris\link{,5}
res <- shrinkda(X,Y)
}
