% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ulda.R
\name{ulda}
\alias{ulda}
\title{uncorrelated linear discriminant analysis}
\usage{
ulda(X, Y, preproc = center(), mu = 0, tol = 1e-06)
}
\arguments{
\item{X}{the data matrix}

\item{Y}{the class labels}

\item{preproc}{the pre-processing function

data(iris)
X <- as.matrix(iris\link{,1:4})
Y <- iris\link{,5}

res <- ulda(X,Y)}
}
\description{
uncorrelated linear discriminant analysis
}
