% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allda.R
\name{update_laplacian}
\alias{update_laplacian}
\title{Update Laplacian Matrix}
\usage{
update_laplacian(S)
}
\arguments{
\item{S}{Similarity matrix (n x n).}
}
\value{
Laplacian matrix L_A (n x n).
}
\description{
Compute the Laplacian matrix L_A from the similarity matrix S.
}
