% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allda.R
\name{update_similarity}
\alias{update_similarity}
\title{Update Similarity Matrix S}
\usage{
update_similarity(X, W, k, r, epsilon = 1e-10)
}
\arguments{
\item{X}{Data matrix of size n x d (n samples, d features).}

\item{W}{Transformation matrix (d x ncomp).}

\item{k}{Number of neighbors.}

\item{r}{Parameter r controlling the influence of distances (should be > 1).}

\item{epsilon}{Small constant to avoid division by zero (default is 1e-10).}
}
\value{
Updated similarity matrix S (n x n).
}
\description{
Update the similarity matrix S using the optimal subspace projection W.
}
