% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allda.R
\name{update_W}
\alias{update_W}
\title{Update Transformation Matrix W}
\usage{
update_W(X, L_A, ncomp, reg = 1e-05)
}
\arguments{
\item{X}{Data matrix of size n x d (n samples, d features).}

\item{L_A}{Laplacian matrix of size n x n.}

\item{ncomp}{Reduced dimension.}

\item{reg}{Regularization term to ensure invertibility of St (default is 1e-5).}
}
\value{
Transformation matrix W (d x ncomp).
}
\description{
Solve for the transformation matrix W using eigen-decomposition.
}
