% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soft_lda.R
\name{weighted_group_means}
\alias{weighted_group_means}
\title{Compute Weighted Group Means}
\usage{
weighted_group_means(X, F)
}
\arguments{
\item{X}{A numeric matrix of size \code{n x d}, where \code{n} is the number of samples and \code{d} is the number of features.}

\item{F}{A numeric matrix of size \code{c x n}, where \code{c} is the number of classes. Each row corresponds to a class,
and each column corresponds to a sample. \code{F[i, j]} is the weight of sample \code{j} for class \code{i}.
Weights must be non-negative.}
}
\value{
A numeric matrix of size \code{c x d}, where each row is the weighted mean of the samples for that class.
}
\description{
Given a data matrix \code{X} and a class weight matrix \code{F}, this function computes the weighted means of the data
for each "class". Unlike hard labels, here each sample contributes partially to each class mean according to its weight.
}
\keyword{internal}
