% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter.R
\name{within_class_scatter}
\alias{within_class_scatter}
\title{Calculate Within-Class Scatter Matrix (via C++)}
\usage{
within_class_scatter(X, Y)
}
\arguments{
\item{X}{A numeric matrix (n x d).}

\item{Y}{A factor or numeric vector of length n.}
}
\value{
A numeric (d x d) within-class scatter matrix.
}
\description{
Computes the within-class scatter matrix using a C++ function \code{withinSScpp}.
}
\keyword{internal}
