% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discursive.R
\name{discursive_combine}
\alias{discursive_combine}
\title{Title}
\usage{
discursive_combine(
  size,
  range,
  constraint,
  type = c("scale", "average", "average_scale", "product")
)
}
\arguments{
\item{constraint}{}
}
\description{
Title
}
