% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discursive_range.R
\name{discursive_range}
\alias{discursive_range}
\title{Compute the range component of discursive sophistication}
\usage{
discursive_range(data, openends)
}
\arguments{
\item{data}{Data frame containing open-ended responses}

\item{openends}{Character vector containing variable names of open-ended items}
}
\value{
Numeric vector of shannon entropy in topic proportions
}
\description{
Compute the range component of discursive sophistication
}
