% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discursive_size.R
\name{discursive_size}
\alias{discursive_size}
\title{Compute the size component of discursive sophistication}
\usage{
discursive_size(
  data,
  openends,
  meta,
  customstopwords = NULL,
  lower.tresh = 10,
  K = 0,
  seed = NULL
)
}
\arguments{
\item{seed}{}
}
\description{
Compute the size component of discursive sophistication
}
