% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discursive_range.R
\name{oe_shannon}
\alias{oe_shannon}
\title{Compute Shannon entropy}
\usage{
oe_shannon(x)
}
\arguments{
\item{x}{Character vector containing open-ended responses}
}
\value{
shannon entropy
}
\description{
Compute Shannon entropy
}
